Saahas bad
to begin this changelog... i need to sleep

regardless I basically rewrote the whole core running of autonomous, instead creating a class this time and due to this allowing easier access to all the variables needed to create a autonomous. In addition this allows for a lot of creation due to lots of prebuilt functions! In addition easy optimization.

Current Beta state contains lots of comments. Easily undoable!